+++ build-html.js
@@ -1,52 +1,101 @@
 /* eslint-env node */
-const fs = require('fs');
-fs.readFile('editor/svg-editor-es.html', 'utf8', (err, data) => {
-  if (err) {
-    console.log('Error reading `svg-editor-es.html` file', err);
-    return;
-  }
-  data = data
-    .replace(
-      '<!DOCTYPE html>',
-      `<!DOCTYPE html>
+const fs = require('promise-fs');
+
+const filesAndReplacements = [
+  {
+    input: 'editor/svg-editor-es.html',
+    output: 'editor/xdomain-svg-editor-es.html',
+    replacements: [
+      [
+        '<script type="module" src="../svgedit-config-es.js"></script>',
+        `<script type="module" src="xdomain-svgedit-config-es.js"></script>`
+      ]
+    ]
+  },
+  {
+    input: 'editor/xdomain-svg-editor-es.html',
+    output: 'editor/xdomain-svg-editor.html',
+    replacements: [
+      [
+        '<!DOCTYPE html>',
+        `<!DOCTYPE html>
+<!-- AUTO-GENERATED FROM xdomain-svg-editor-es.html; DO NOT EDIT; use build-html.js to build -->`
+      ],
+      [
+        '<script type="module" src="redirect-on-lacking-support.js"></script>',
+        '<script defer="defer" src="../dist/redirect-on-lacking-support.js"></script>'
+      ],
+      [
+        '<script type="module" src="xdomain-svgedit-config-es.js"></script>',
+        '<script defer="defer" src="xdomain-svgedit-config-iife.js"></script>'
+      ],
+      [
+        '<script src="external/dom-polyfill/dom-polyfill.js"></script>',
+        '<script src="../dist/dom-polyfill.js"></script>'
+      ]
+    ]
+  },
+  {
+    input: 'editor/svg-editor-es.html',
+    output: 'editor/svg-editor.html',
+    replacements: [
+      [
+        '<!DOCTYPE html>',
+        `<!DOCTYPE html>
 <!-- AUTO-GENERATED FROM svg-editor-es.html; DO NOT EDIT; use build-html.js to build -->`
-    ).replace(
-      '<script type="module" src="redirect-on-lacking-support.js"></script>',
-      '<script defer="defer" src="../dist/redirect-on-lacking-support.js"></script>'
-    ).replace(
-      '<script type="module" src="../svgedit-config-es.js"></script>',
-      '<script defer="defer" src="../svgedit-config-iife.js"></script>'
-    ).replace(
-      '<script src="external/dom-polyfill/dom-polyfill.js"></script>',
-      '<script src="../dist/dom-polyfill.js"></script>'
-    );
-  fs.writeFile('editor/svg-editor.html', data, (err) => {
-    if (err) {
-      console.log('Error writing file', err);
-      return;
-    }
-    console.log('Completed file rewriting!');
-  });
-});
-fs.readFile('editor/extensions/imagelib/index-es.html', 'utf8', (err, data) => {
-  if (err) {
-    console.log('Error reading `imagelib/index-es.html` file', err);
-    return;
-  }
-  data = data
-    .replace(
-      '<!DOCTYPE html>',
-      `<!DOCTYPE html>
+      ],
+      [
+        '<script type="module" src="redirect-on-lacking-support.js"></script>',
+        '<script defer="defer" src="../dist/redirect-on-lacking-support.js"></script>'
+      ],
+      [
+        '<script type="module" src="../svgedit-config-es.js"></script>',
+        '<script defer="defer" src="../svgedit-config-iife.js"></script>'
+      ],
+      [
+        '<script src="external/dom-polyfill/dom-polyfill.js"></script>',
+        '<script src="../dist/dom-polyfill.js"></script>'
+      ]
+    ]
+  },
+  {
+    input: 'editor/extensions/imagelib/index-es.html',
+    output: 'editor/extensions/imagelib/index.html',
+    replacements: [
+      [
+        '<!DOCTYPE html>',
+        `<!DOCTYPE html>
 <!-- AUTO-GENERATED FROM imagelib/index-es.html; DO NOT EDIT; use build-html.js to build -->`
-    ).replace(
-      '<script type="module" src="index.js"></script>',
-      '<script defer="defer" src="../../../dist/extensions/imagelib/index.js"></script>'
-    );
-  fs.writeFile('editor/extensions/imagelib/index.html', data, (err) => {
-    if (err) {
-      console.log('Error writing file', err);
+      ],
+      [
+        '<script type="module" src="index.js"></script>',
+        '<script defer="defer" src="../../../dist/extensions/imagelib/index.js"></script>'
+      ]
+    ]
+  }
+];
+
+filesAndReplacements.reduce((p, {input, output, replacements}) => {
+  return p.then(async () => {
+    let data;
+    try {
+      data = await fs.readFile(input, 'utf8');
+    } catch (err) {
+      console.log(`Error reading ${input} file`, err);
+    }
+
+    data = replacements.reduce((s, [find, replacement]) => {
+      return s.replace(find, replacement);
+    }, data);
+
+    try {
+      await fs.writeFile(output, data);
+    } catch (err) {
+      console.log(`Error writing file: ${err}`, err);
       return;
     }
-    console.log('Completed file rewriting!');
+    console.log(`Completed file ${input} rewriting!`);
   });
+}, Promise.resolve()).then(() => {
+  console.log('Finished!');
 });