+++ editor/xdomain-svgedit-config-es.js
@@ -1,0 +1,7 @@
+import svgEditor from './svg-editor.js';
+svgEditor.setConfig({
+  allowedOrigins: ['*']
+});
+
+// <CONDITIONAL-ADD>: window.svgEditor = svgEditor; // DO NOT DELETE THIS LINE: USED BY AUTOMATED SCRIPT
+// <CONDITIONAL-ADD>: window.svgEditor.modules = false; // DO NOT DELETE THIS LINE: USED BY AUTOMATED SCRIPT